22-4-22_2

+++ review_management/app/controllers/review.controller.js
@@ -1,55 +1,52 @@
 require("dotenv").config()
 const reviewService = require("../services/review.service")
-const review2 = require("../models/review")
 const rabbitClient = require("../config/reviewdb")
 
 const url = process.env.RABBIT_MQ_URL
 
 module.exports = {
     getAllReviews: async (req, res) => {
-        const review = await reviewService.getAllReviews()
-        res.statusCode = 200
-        res.setHeader('Content-Type', 'application/json')
-        res.json(review)
+        const reviews = await reviewService.getAllReviews()
+        if (reviews.StatusCode == null)
+            return res.status(200).json(reviews);
+        else
+            res.status(reviews.StatusCode).json(reviews);
     },
     getReview: async (req, res) => {
-        const review = await reviewService.getReview(req.params.review_Id)
-        if (review.statusCode == 200 && (res.statusCode >= 200 && res.statusCode < 400)) {
-            res.statusCode = 200
-            res.setHeader('Content-Type', 'application/json')
-            res.json(review)
-        }
-        else {
-            res.statusCode = 400
-            res.setHeader('Content-Type', 'application/json')
-            res.json(review)
-        }
+        const reviews = await reviewService.getReview(req.params.review_Id)
+        if (reviews.StatusCode == null)
+            return res.status(200).json(reviews);
+        else
+            res.status(reviews.StatusCode).json(reviews);
     },
     postReview: async (req, res) => {
         try {
-            const review = await reviewService.postReview(req.body)
-            if (review.statusCode != 400) {
+            const checkValid = await reviewService.checkOrderValid(req.params.order_Id)
+            if (checkValid.acknowledged) {
+                const review = await reviewService.postReview(req.body)
+                if (review.statusCode != 400) {
 
-                const avgRating = await reviewService.checkAvgRating(req.body.restaurant_Id)
-                var avgRatingField = avgRating[0].AverageValue.toFixed(1);
-                const totalRating = await reviewService.checktotalRatings(req.body.restaurant_Id)
-                var totalRatingField = totalRating[0].TotalRatings;
-                rabbitClient.client.connect(url, function (err, conn) {
-                    if (err != null) bail(err);
-                    console.log("Connected , Publish Review")
-                    const data = {
-                        restaurant_id: req.body.restaurant_Id,
-                        avg_rating: avgRatingField,
-                        totalRatings: totalRatingField
-                    }
-                    rabbitClient.publish_review(conn, data);
-                    console.log("Review Published")
-                });
+                    const avgRating = await reviewService.checkAvgRating(req.body.review_Id)
+                    var avgRatingField = avgRating[0].AverageValue.toFixed(1);
+                    const totalRating = await reviewService.checktotalRatings(req.body.review_Id)
+                    var totalRatingField = totalRating[0].TotalRatings;
+                    rabbitClient.client.connect(url, function (err, conn) {
+                        if (err != null) bail(err);
+                        console.log("Connected , Publish Review")
+                        const data = {
+                            restaurant_id: req.body.restaurant_Id,
+                            avg_rating: avgRatingField,
+                            totalRatings: totalRatingField
+                        }
+                        rabbitClient.publish_review(conn, data);
+                        console.log("Review Published")
+                    });
 
-                res.status(201).json(review);
-            }
-            else {
-                res.status(400).json(review);
+                    res.status(201).json(review);
+                }
+                else {
+                    res.status(400).json(review);
+                }
             }
         }
         catch (e) {
@@ -61,34 +58,24 @@ module.exports = {
         }
     },
     deleteReview: async (req, res) => {
-        const check = await reviewService.deleteReview(req.params.review_Id)
-        if (check.acknowledged) {
-            res.statusCode = 200
-            res.setHeader('Content-Type', 'application/json')
-            res.json(check)
-        }
-        else {
-            res.statusCode = 400
-            res.setHeader('Content-Type', 'application/json')
-            res.json(check)
+        const checkValid = await reviewService.checkOrderValid(req.params.order_Id)
+        if (checkValid.acknowledged) {
+            const check = await reviewService.deleteReview(req.params.review_Id)
+            if (check.acknowledged) {
+                return res.status(200).json(check);
+            }
+            else {
+                return res.status(check.StatusCode).json(check);
+            }
         }
     },
     updateReview: async (req, res) => {
-        try {
-            const review = await reviewService.updateReview(req.params.id, req.body)
-            if (review.statusCode != 400) {
-                res.status(201).json(review);
-            }
-            else {
-                res.status(400).json(review);
-            }
+        const reviews = await reviewService.updateReview(req.params.review_Id, req.body)
+        if (reviews.StatusCode == null) {
+            return res.status(200).json(reviews);
         }
-        catch (e) {
-            if (!e.status) {
-                res.status(500).json({ error: { code: 'UNKNOWN_ERROR', message: 'An unknown error occurred.' } });
-            } else {
-                res.status(e.status).json({ error: { code: e.code, message: e.message } });
-            }
+        else {
+            return res.status(reviews.StatusCode).json(reviews);
         }
     }
 }
